Move some utility functions into a new util.py file.
Add support for an Action message in the log which represents when action was taken to change a phone's forwarding state.

Change the home phone script so that it will toggle phone status (and log that it changed) for any phone change that doesn't have a corresponding log entry showing that action was taken.

Wire the --enable and --disable flag so that they add actions if no pending actions need to happen to ensure there is a log action for every change action in the log.