Oops, I did this backwards.  When a phone comes home that means it should be disabled and it should be enabled when it leaves.
The whole point is that we have an OBi100 at home for the Google Voice number and it's redundant to have both the OBi100 ring _and_ the cell phones ring.  Especially because the OBi100 is plugged into a panasonic link-to-cell cordless phoone which is paired via bluetooth to our cell phones.